Use size_t rather than uint32 in a few more places to fixup some warnings.
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/274009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28747 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 82809538ae7150d40c8e0edc9020915f46f6c974

+++ net/flip/flip_framer.cc
@@ -249,7 +249,7 @@ size_t FlipFramer::ProcessInput(const char* data, size_t len) {
               if (!InitializeDecompressor())
                 return NULL;
 
-              int decompressed_max_size = amount_to_forward * 100;
+              size_t decompressed_max_size = amount_to_forward * 100;
               scoped_array<char> decompressed(new char[decompressed_max_size]);
               decompressor_->next_in = reinterpret_cast<Bytef*>(
                   const_cast<char*>(data));
@@ -263,7 +263,7 @@ size_t FlipFramer::ProcessInput(const char* data, size_t len) {
                 set_error(FLIP_DECOMPRESS_FAILURE);
                 goto bottom;
               }
-              int decompressed_size = decompressed_max_size -
+              size_t decompressed_size = decompressed_max_size -
                                       decompressor_->avail_out;
               visitor_->OnStreamFrameData(current_data_frame->stream_id(),
                                           decompressed.get(),
